+++ pyCADD/Dance/__init__.py
@@ -1,4 +1,4 @@
 '''
 Data Analyzer for Computer-aid drug design
 '''
-from pyCADD.Dance.core import Matrix, Evaluator
+from pyCADD.Dance.common import Matrix, Dancer, Evaluator
